+++ InteliJProject/src/SpeedConfiguration.java
@@ -4,75 +4,75 @@ public class SpeedConfiguration {
 	/**
 	 * Speed for perimeters (mm/s)
 	 */
-	private int perimetersSpeed;
+	private double perimetersSpeed;
 	
 	/**
 	 * Speed for small perimeters (mm/s)
 	 */
-	private int smallPerimetersSpeed;
+	private double smallPerimetersSpeed;
 	
 	/**
 	 * Speed for external perimeters (mm/s)
 	 */
-	private int externalPerimetersSpeed;
+	private double externalPerimetersSpeed;
 	
 	/**
 	 * Speed for infill (mm/s)
 	 */
-	private int infillSpeed;
+	private double infillSpeed;
 	
 	/**
 	 * Speed for solid infill (mm/s)
 	 */
-	private int solidInfillSpeed;
+	private double solidInfillSpeed;
 	
 	/**
 	 * Speed for top solid infil (mm/s)
 	 */
-	private int topSolidInfillSpeed;
+	private double topSolidInfillSpeed;
 	
 	/**
 	 * Speed for support material (mm/s)
 	 */
-	private int supportMaterialSpeed;
+	private double supportMaterialSpeed;
 	
 	/**
 	 * Speed for bridges (mm/s)
 	 */
-	private int bridgesSpeed;
+	private double bridgesSpeed;
 	
 	/**
 	 * Speed for gap fill (mm/s)
 	 */
-	private int gapFillSpeed;
+	private double gapFillSpeed;
 	
 	/**
-	 * Speed for non print movements (mm/s)
+	 * Speed for non prdouble movements (mm/s)
 	 */
-	private int nonPrintMovesSpeed;
+	private double nonPrintMovesSpeed;
 	
 	/**
 	 * Speed for the first layer (mm/s)
 	 */
-	private int firstLayerSpeed;
+	private double firstLayerSpeed;
 	
 	/**
 	 * Acceleration for perimeters (mm/s^2)
 	 */
-	private int perimetersAcceleration;
+	private double perimetersAcceleration;
 	
 	/**
 	 * Acceleration for infill (mm/s^2)
 	 */
-	private int infillAcceleration;
+	private double infillAcceleration;
 	
 	/**
 	 * Acceleration for bridges (mm/s^2)
 	 */
-	private int bridgeAcceleration;
+	private double bridgeAcceleration;
 	
 	/**
 	 * Default Acceleration (mm/s^2)
 	 */
-	private int defaultAcceleration;
+	private double defaultAcceleration;
 }
